ci: AddressSanitizer/UBSan and clang-tidy, plus the fixes they needed
Adds a Linux LLVM ASan+UBSan preset (cmake/Sanitizers.cmake,
config-linux-llvm-asan-ubsan) with -fno-sanitize-recover=all so a violation
fails the build outright rather than logging and continuing, and a curated
.clang-tidy (bugprone-*, clang-analyzer-*, performance-*, a narrow cert-*
subset). Both get a CI leg: the sanitizer job runs the full test suite plus
a new scripts/run-codec-matrix.sh - ctest's cases exercise codec logic in
isolation, and a sanitizer only catches what actually executes, so the
script drives the real CLI across every layout, every Annex E tool token,
and both Atmos modes. Deliberately excludes -fsanitize=integer's unsigned-
overflow check (CRC16, ring buffers and the splice counter rely on
intentional wraparound) and two clang-tidy checks verified to be false
positives here (bugprone-easily-swappable-parameters is pure API-design
advice; the two widening-cast checks flagged 34 sites that all trace to
spec-bounded bitfields or codec dimensions nowhere near overflow range).
cppcheck was evaluated and not adopted as a gate - it found nothing beyond
one redundant condition, fixed here, under its full check set.

Getting Linux Clang/GCC to compile at all first needed five small
-Wdouble-promotion/-Wimplicit-int-float-conversion casts and, on GCC, a
-Wshadow collision between chanmap's bitmask constants and its Location
enum (renamed the constants with a Bit suffix), a range-loop copy, and a
sign-conversion site in wav.cpp. Both legs are proven green now (214/214
tests, verified in an Ubuntu 26.04 container matching CI's own toolchain
scripts) and promoted out of `experimental: true`; the CI comment blocks
claiming otherwise were stale.

Real findings the tools caught: a redundant condition in mixing.cpp
(cppcheck), a missing views.reserve() and two blocked moves in main.cpp,
and ten small error enums given an explicit uint8_t underlying type
(performance-enum-size). main() gained a top-level try/catch, since nothing
previously stopped a std::format_error from escaping to a bare terminate.
Six clang-tidy findings are NOLINT-suppressed with one-line justifications
where verified safe by construction - optional accesses proven engaged by
control flow, a static-init that cannot throw, a divide that cannot be
zero across the whole call graph even though the analyzer cannot see that
across translation units.

Also extends .github/toolchain/03-llvm-toolchain.sh with the sanitizer
runtime and symbolizer packages (libclang-rt-21-dev, llvm-21) neither
Linux LLVM leg needed until now.

Not fixed: the E-AC-3 decoder refuses any stream using coupling, spectral
extension or AHT (eac3_decoder.cpp returns kUnsupported the moment cplinu/
spxinu/ahte is set), even though eac3-encode happily produces one -
tools/quality_race.py already routes tool-enabled streams through FFmpeg
instead of this decoder, suggesting a deliberate scope boundary rather
than a bug, but that's unconfirmed. Recorded rather than guessed at.

+++ src/lib/src/encoder/plan.cpp
@@ -862,7 +862,7 @@ void render(const Routing& routing, std::span<const std::span<const float>> sour
             const auto in = source[static_cast<std::size_t>(s)];
             const auto n = std::min(samples, in.size());
             for (std::size_t i = 0; i < n; ++i) {
-                out[i] += static_cast<float>(gain * in[i]);
+                out[i] += static_cast<float>(gain * static_cast<double>(in[i]));
             }
         }
     }
